# add optimized implementation for AVX512 targets