--- a/xxhsum.c
+++ b/xxhsum.c
@@ -387,7 +387,9 @@ static unsigned BMK_isLittleEndian(void)
 
 /* Try to detect the architecture. */
 #if defined(ARCH_X86)
-#  if defined(__AVX2__)
+#  if defined(__AVX512F__)
+#    define ARCH ARCH_X86 " + AVX512"
+#  elif defined(__AVX2__)
 #    define ARCH ARCH_X86 " + AVX2"
 #  elif defined(__AVX__)
 #    define ARCH ARCH_X86 " + AVX"