@easyaspi314 2020-04-06
> on avx512 targets, variant with default secrets will improve 45%. The following result is compiled with `-Ofast -mavx512`, and there is also no need to add `-mno-avx256-split-unaligned-load` with gcc.

Those benchmarks are made with Clang, which likes to impersonate `gcc` and MSVC. 

Notably, FreeBSD, macOS, the Android NDK, and Termux all do this by default.