@gzm55 2020-04-06
the previous benchmarks are compiled via `CC=clang-9 make` on a Ubuntu 18.04 machine, the GCC compiled version is also tested with a similar result.

For the `-mno-avx256-split-unaligned-load` option, the real GCC is not affected any more since that is an `avx2 only` option.