@gzm55 2020-04-06
The benchmark results compiled by gcc is updated.

It is wired that custom secret will significantly affects performance, the clang-compiled avx512 `XXH3_64b w/secret` even slower than avx2.