@easyaspi314 2020-04-06
Clang doesn't appear to even *try* to unroll the `withSecret` variants.