@gzm55 2020-04-06
gcc also has the similar performance on the `withSecret` ones.