@easyaspi314 2020-04-06
What about this?

```c
#if XXH_VECTOR == XXH_AVX512 && !XXH_REROLL
    if (XXH_likely(nbStripes % 4 == 0)) {
        for (n = 0; n < nbStripes; n+=4 ) {
            const xxh_u8* const in = input + n*STRIPE_LEN;
            if (accWidth == XXH3_acc_64bits) XXH_PREFETCH(in + XXH_PREFETCH_DIST_AVX512_64);
            else                             XXH_PREFETCH(in + XXH_PREFETCH_DIST_AVX512_128);
            XXH3_accumulate_512(acc,
                                in,
                                secret + n*XXH_SECRET_CONSUME_RATE,
                                accWidth);
            XXH3_accumulate_512(acc,
                                in + 1 * STRIPE_LEN,
                                secret + (n+1)*XXH_SECRET_CONSUME_RATE,
                                accWidth);
            XXH3_accumulate_512(acc,
                                in + 2 * STRIPE_LEN,
                                secret + (n+2)*XXH_SECRET_CONSUME_RATE,
                                accWidth);
            XXH3_accumulate_512(acc,
                                in + 3 * STRIPE_LEN,
                                secret + (n+3)*XXH_SECRET_CONSUME_RATE,
                                accWidth);
        }
    } else
#endif
    for (n = 0; n < nbStripes; n++) {
        ...
```