@Cyan4973 2020-04-06
Thanks for this great submission @gzm55 !

I understand it noticeably improves speed when using the default `secret` for CPU with AVX512 instruction set. +40% definitely matches the bar in order to introduce a new target-specific code path in the library.

I also understand that optimizing the `_withSecret` variant is more complex.
That being said, as a way to keep this PR clean and simple, I'm happy to merge it if you wish, and accept a follow-up with improvements on the `_withSecret` variant.

I merely have a small question on the new prefetch distance.