@Cyan4973 2020-04-06
> Clang doesn't appear to even try to unroll the withSecret variants, so there is a lot of branching overhead.

Was it tested with `XXH_INLINE_ALL` ?
I naively assumed that, if the `secret` is known at compile time, then the compiler would find a way to use the `secret`'s length to correctly optimize the inner loop, even if it's not the default one.