@gzm55 2020-04-06
@Cyan4973 agree to separate the `_withSecret` discussion and optimization to another PR.

the distance of prefetching is got from profiling each one byte for 100K length input (xxhsum default). The benchmark machine is a ECS server, not a desktop. Do you have some experience or environments to settle down a better distance?