@Cyan4973 2020-04-06
> Do you have some experience or environments to settle down a better distance?

There will never be a "one size fits all" best prefetch distance for all systems.
The "better" distance is necessarily system dependent.

Some generic rules of thumb to establish a distance : 
- Currently, this is a "better than nothing" approach; we want a "generic" value which is good enough to improve performance across the board. This is different from looking for the _best_ value for each system.
- In the future, we might have to offer some programmability on this value. First compile time, then maybe runtime. We don't have to do that now, this is ground for some future patch.
- Be wary of synthetic benchmark for this topic. A benchmark will basically assume that it has the whole system for itself. This will lead to over-use of memory bandwidth and cache size, on the ground that they improve the synthetic benchmark. But in a real application, bandwidth will be shared with other parts of the application, and over-using it with over-demanding prefetch requests will lead to system-wide performance degradation. When selecting a prefetching distance, one wants to lean towards _smaller_ distances, and basically stop early, as soon as benefits from larger prefetching distances become "small". That's because it preserve a general resource (memory bandwidth) for the benefit of the larger system, be it other parts of the application or other concurrent applications.